moved adc calibration functions to their own file
