+++ app_grim_reefer/prj.conf
@@ -32,9 +32,10 @@ CONFIG_LSM6DSL_ACCEL_ODR=8
 CONFIG_LSM6DSL_GYRO_ODR=8
 
 # 16g
-LSM6DSL_ACCEL_FS=16
+
+CONFIG_LSM6DSL_ACCEL_FS=16
 # 2000 degrees per second
-LSM6DSL_GYRO_FS=2000
+CONFIG_LSM6DSL_GYRO_FS=2000
 
 
 # Serial for data dumping
