+++ app_grim_reefer/src/testing.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <zephyr/drivers/adc.h>
+
+void adc_printout(const struct adc_dt_spec *channel);
